+++ python/GafferImageTest/MedianTest.py
@@ -274,8 +274,8 @@ class MedianTest( GafferImageTest.ImageTestCase ) :
 
 		driverShuffle = GafferImage.Shuffle()
 		driverShuffle["in"].setInput( imageReader["out"] )
-		driverShuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "driver", "R" ) )
-		driverShuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "driven", "R" ) )
+		driverShuffle["shuffles"].addChild( Gaffer.ShufflePlug( "R", "driver" ) )
+		driverShuffle["shuffles"].addChild( Gaffer.ShufflePlug( "R", "driven" ) )
 
 		driverDelete = GafferImage.DeleteChannels()
 		driverDelete["in"].setInput( driverShuffle["out"] )
@@ -287,8 +287,8 @@ class MedianTest( GafferImageTest.ImageTestCase ) :
 
 		refDriverShuffle = GafferImage.Shuffle()
 		refDriverShuffle["in"].setInput( refReader["out"] )
-		refDriverShuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "driver", "R" ) )
-		refDriverShuffle["channels"].addChild( GafferImage.Shuffle.ChannelPlug( "driven", "R" ) )
+		refDriverShuffle["shuffles"].addChild( Gaffer.ShufflePlug( "R", "driver" ) )
+		refDriverShuffle["shuffles"].addChild( Gaffer.ShufflePlug( "R", "driven" ) )
 
 		refDriverDelete = GafferImage.DeleteChannels()
 		refDriverDelete["in"].setInput( refDriverShuffle["out"] )
@@ -331,7 +331,7 @@ class MedianTest( GafferImageTest.ImageTestCase ) :
 					continue
 
 				with self.subTest( refFile = ref, driverChannel = channelName ):
-					for channelPlug in [ i["in"] for i in driverShuffle["channels"].children() + refDriverShuffle["channels"].children() ] + [ driverMedian["masterChannel"] ]:
+					for channelPlug in [ i["source"] for i in driverShuffle["shuffles"].children() + refDriverShuffle["shuffles"].children() ] + [ driverMedian["masterChannel"] ]:
 						channelPlug.setValue( channelName )
 					self.assertImagesEqual( driverMedian["out"], refDriverDelete["out"], ignoreMetadata = True )